splitting out travel domain
